memcpy

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: erijania <erijania@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/20 19:22:46 by erijania          #+#    #+#             */
-/*   Updated: 2024/02/21 09:03:53 by erijania         ###   ########.fr       */
+/*   Updated: 2024/02/22 10:14:39 by erijania         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*tmp;
